web/pyproject.toml
[project]
name = "web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.4,<6.0",
    "django-celery-beat>=2.8.1",
    "django-celery-results>=2.6.0",
    "django-htmx>=1.23.2",
    "htpy>=25.12.0",
    "httpx>=0.27.2",
    "pyjwt>=2.9.0",
    "redis>=6.0.0",
    "sourcetypes>=0.0.4",
]

[tool.htpy-uikit]
components_dir = "components/ui"
theme_path = "staticfiles/css/theme.css"

[dependency-groups]
dev = [
    "django-stubs>=5.2.2",
    "django-stubs-ext>=5.2.1",
    "pyright>=1.1.408",
    "ruff>=0.15.0",
]

[tool.pyright]
typeCheckingMode = "basic"
reportAttributeAccessIssue = false
reportIncompatibleVariableOverride = false
include = ["web", "core"]
exclude = ["**/migrations/**", "components/**", "staticfiles/**"]
